syscall: int arguments were judged as whole 64-bit registers
A syscall argument declared int or unsigned int reaches the kernel as
the low 32 bits of its register; whatever the high half holds is not
part of the call. A handful of validations here read the whole
register instead, so a value the kernel accepts was refused:

  * mlock2's flags (int): a high bit was EINVAL.
  * getgroups' gidsetsize (int): a count of 0 with high bits set was
    judged too small instead of answering the group count, and a
    negative one was not EINVAL up front the way the kernel has it.
  * seccomp's op and flags (unsigned int): the switch on the whole
    register sent a valid op with high bits to the EINVAL default, and
    GET_ACTION_AVAIL refused flags that were all in the high half.
  * sched_getaffinity / sched_setaffinity's len (unsigned int): a high
    half made a legal length look enormous, so the whole 4 KiB mask
    buffer went to the host in place of the eight bytes asked for.

Each now narrows first. A probe built for the host and for the guest
agrees on every case, in both engines and on the ILP32 build.

+++ src/sys_mm.c
@@ -1080,9 +1080,10 @@ SYSDEF(mincore) {
 SYSDEF(mlock) { (void)c;(void)a0;(void)a1;(void)a2;(void)a3;(void)a4;(void)a5; return 0; }
 SYSDEF(mlock2) {
     /* Accept-and-ignore like the rest of the mlock family, but keep the
-     * kernel's flag validation: only MLOCK_ONFAULT (1) is defined. */
+     * kernel's flag validation: only MLOCK_ONFAULT (1) is defined. The flags
+     * are an int, so the register's high half is not part of them. */
     (void)c; (void)a0; (void)a1; (void)a3; (void)a4; (void)a5;
-    return (a2 & ~1ULL) ? (u64)(s64)-EINVAL : 0;
+    return ((unsigned)a2 & ~1u) ? (u64)(s64)-EINVAL : 0;
 }
 SYSDEF(munlock) { (void)c;(void)a0;(void)a1;(void)a2;(void)a3;(void)a4;(void)a5; return 0; }
 SYSDEF(mlockall) { (void)c;(void)a0;(void)a1;(void)a2;(void)a3;(void)a4;(void)a5; return 0; }
